refactor: Separate queue service and implement factory for webhooks
