+++ src/main/java/com/lpvs/entity/report/package-info.java
@@ -1,0 +1,11 @@
+/**
+ * Copyright (c) 2024, Samsung Electronics Co., Ltd. All rights reserved.
+ *
+ * Use of this source code is governed by a MIT license that can be
+ * found in the LICENSE file.
+ */
+
+/**
+ * This package contains entity classes that represent various types of reports generated by the LPVS service.
+ */
+package com.lpvs.entity.report;
