+++ src/main/java/com/lpvs/service/webhook/package-info.java
@@ -1,0 +1,11 @@
+/**
+ * Copyright (c) 2024, Samsung Electronics Co., Ltd. All rights reserved.
+ *
+ * Use of this source code is governed by a MIT license that can be
+ * found in the LICENSE file.
+ */
+
+/**
+ * The package contains classes related to the LPVS Webhook Service operation.
+ */
+package com.lpvs.service.webhook;
